kvutils: Extract out a `WriteSetSelector`.
We want to replace this with one that's always successful in the
integrity checker.
